+++ tsante-mqtt-publisher.js
@@ -6,7 +6,7 @@ Polymer({
      * the topic to publish in
      *
      * when the topic is changed, the payload is changed to null.
-     * 
+     *
      * @type {String}
      */
     topic: {
@@ -39,23 +39,38 @@ Polymer({
       type: String,
       value: null,
     },
+    /**
+     * the connection status of the `tsante-mqtt` ancestor
+     * @type {Boolean}
+     */
+    _connected: {
+      type: Boolean,
+      value: false,
+    },
+    /**
+     * the mqtt client of the `tsante-mqtt` ancestor
+     * @type {Object}
+     */
+    _client: Object
   },
 
   observers: [
+    '_isConnected(_connected)',
     'publish(payload)',
   ],
 
-  attached: function() {
-    if (this.parentElement.tagName !== 'tsante-mqtt'.toUpperCase()) {
-      console.error('tsante-mqtt-publisher must have a tsante-mqtt parent');
-    }
-    this.parentElement.addEventListener('tsante-mqtt-connect', (evt) => {
-      if(evt.detail.status && this.payload) { this.publish(); }
-    });
+  setNeededProperties: function(connected, client){
+    this._connected = connected
+    this._client = client
+  },
+
+  _isConnected: function(_connected){
+    if (_connected  && this._client) this.publish();
   },
 
   _topicChanged: function(newValue) {
     this.payload = null;
+    this.alreadyPublished = false;
   },
 
   /**
@@ -67,12 +82,12 @@ Polymer({
    * @param  {Boolean} retained flag indicates that the server must or not keep the value by default this.retained
    */
   publish: function(payload, qos, retained) {
-      if(this.parentElement.connected) {
-        payload = typeof payload === "string"?payload:this.payload;
-        if(payload === null) return;
-        qos = (!qos || isNaN(qos))?this.qos:qos;
-        retained = typeof retained === 'boolean'?retained:this.retained;
-        this.parentElement.client.send(this.topic, payload, qos, retained);
-      }
+    if(this._connected) {
+      payload = typeof payload === 'string'?payload:this.payload;
+      if(payload === null) return;
+      qos = (!qos || isNaN(qos))?this.qos:qos;
+      retained = typeof retained === 'boolean'?retained:this.retained;
+      this._client.send(this.topic, payload, qos, retained);
+    }
   }
 });