+++ tsante-mqtt.js
@@ -6,7 +6,7 @@ Polymer({
     /**
      * the url of the server
      *
-     * by exapmple :
+     * by example :
      * `ws://test.mosquitto.org:8080/`
      *
      * @type {String}
@@ -109,9 +109,26 @@ Polymer({
   },
 
   observers: [
+    '_sendInfoToPubSub(connected, client)',
     '_initializeClient(host,clientID)',
     'connect(username, password)',
   ],
+
+
+  /**
+   * on connected change or client change
+   * update publishers and subscribers children
+   * @method sendInfoToPubSub
+   *
+   * @param {*} connected
+   * @param {*} client
+   */
+  _sendInfoToPubSub(connected, client) {
+    let pubSubs = [].slice.call(this.querySelectorAll('tsante-mqtt-publisher, tsante-mqtt-subscriber'))
+    for(let pubSub of pubSubs){
+      pubSub.setNeededProperties(connected, client)
+    }
+  },
 
   attached() {
     this.connect();
@@ -166,7 +183,19 @@ Polymer({
    * @param {Boolean} status the status of the connection
    */
   _connectedChanged: function() {
-    this.fire('tsante-mqtt-connect', { status: this.connected });
+    let subscribers = [].slice.call(this.querySelectorAll('tsante-mqtt-subscriber'));
+    if(this.connected) {
+      this.fire('tsante-mqtt-connect', { status: this.connected });
+    }
+    for(let subscriber of subscribers){
+      if(this.connected) {
+        subscriber.subscribe(this.connected, this.client)
+        subscriber._setSubscribed(true)
+      // }else{
+        // value._setSubscribed(false);
+        // value.fire('tsante-mqtt-subscribed', { topic: this.topic, status: false });
+      }
+    }
   },
 
   _onConnectionLost: function(msg) {
@@ -175,23 +204,14 @@ Polymer({
   },
 
   /**
-   * fired when a message is read
+   * when a message is received transfert it to the subscribers
    *
-   * example of the `evt.detail` :
-   *
-   * ```
-   * {
-   *   topic: "terminal/hello",
-   *   payload: "polymer"
-   * }
-   * ```
-   *
-   * @event tsante-mqtt-received
-   * @param  {String} topic topic of the received message
-   * @param  {String} payload content of the received message
    */
   _onMessageArrived: function(msg) {
-    this.fire('tsante-mqtt-received', { topic:msg.destinationName, payload:msg.payloadString })
+    let subscribers = [].slice.call(this.querySelectorAll('tsante-mqtt-subscriber'))
+    for(let subscriber of subscribers){
+      subscriber._received(msg);
+    }
   },
 
   /**
@@ -217,7 +237,9 @@ Polymer({
      * @param {String} topic the topic
      * @param {String} payload the content of the message
      */
-    this.fire('tsante-mqtt-delivered', { topic:msg.destinationName, payload:msg.payloadString })
+    if(typeof msg.payloadString === 'string'){
+      this.fire('tsante-mqtt-delivered', { topic:msg.destinationName, payload:msg.payloadString })
+    }
   },
 
   /**
@@ -237,12 +259,13 @@ Polymer({
         cleanSession: this.cleanSession,
       };
       if(this.username || username) {
+        this.username = this.username || username;
         connectOption.userName = this.username || username;
       }
       if(this.password || password) {
+        this.password = this.password || password;
         connectOption.password = this.password || password;
       }
-
       try {
         this.client.connect(connectOption);
       } catch(err) {
@@ -256,9 +279,11 @@ Polymer({
    * @method disconnect
    */
   disconnect: function() {
+    // if (!this.client) return;
     try {
       this.client.disconnect();
       this._setConnected(false);
+      this.fire('tsante-mqtt-connect', { status: this.connected });
     } catch(err) {
       this._onError(err.message);
     }